chromium: fix v4l2 compiler error on arm
fixes:
../../media/gpu/v4l2/legacy/v4l2_video_decoder_backend_stateful.cc:450:20:
error: non-constant-expression cannot be narrowed from type '__suseconds64_t'
(aka 'long long') to 'long' in initializer list [-Wc++11-narrowing]

Upstream-Status: Inappropriate [specific to our build setup]

+++ src/media/gpu/v4l2/legacy/v4l2_video_decoder_backend_stateful.cc
@@ -440,7 +440,7 @@ void V4L2StatefulVideoDecoderBackend::OnOutputBufferDequeued(
     const struct timeval timeval = buffer->GetTimeStamp();
     const struct timespec timespec = {
         .tv_sec = timeval.tv_sec,
-        .tv_nsec = timeval.tv_usec * 1000,
+        .tv_nsec = static_cast<long>(timeval.tv_usec * 1000),
     };
 
     const int64_t flat_timespec =
